% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prithulib.R
\name{cards}
\alias{cards}
\title{Choosing a card}
\usage{
cards(n, joker = FALSE)
}
\arguments{
\item{n}{is the number of cards to be chosen (default 1)}

\item{joker}{TRUE includes the Joker card in the deck (default FALSE)}
}
\value{
The chosen card(s)
}
\description{
An event of choosing a card from a pack of 52 cards (Joker optional)
}
\examples{
cards(3, joker = TRUE)
}
\author{
Prithul Chaturvedi
}
