% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prithulib.R
\name{roll}
\alias{roll}
\title{Rolling an unbiased die}
\usage{
roll(n = 1, size = 6)
}
\arguments{
\item{n}{is the number of times to roll the die (default 1)}

\item{size}{is the number of faces on the die (default 6)}
}
\value{
Outcome(s) of rolling of the die
}
\description{
Rolls a die for 'n' number of times
}
\examples{
roll(2,10)
}
\author{
Prithul Chaturvedi
}
