% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prithulib.R
\name{toss}
\alias{toss}
\title{Tossing an unbiased coin}
\usage{
toss(n = 1)
}
\arguments{
\item{n}{is the number of times the coin is tossed (default 1)}
}
\value{
Outcome(s) of tossing the coin
}
\description{
Tosses a coin for 'n' number of times
}
\examples{
toss(3)
}
\author{
Prithul Chaturvedi
}
